Fixed a bug - File system errors if file size is exactly Cluster size.
BUG
We get the next write sector even if we don't need it.
This means that we can allocate a whole cluster we don't use.
That results in filesystem errors of cluster chain longer than filesize.

FIX
The fix is that we now check if we need a new sector before we go get one.
Fixed in both InitBlob and FileWrite